+++ server/index.prod.js
@@ -223,7 +223,7 @@ export const createServer = (config) => {
                 <script src="${assets.vendor.js}"></script>
                 <script async src="${assets.main.js}" ></script>
                 <link rel='stylesheet' type='text/css' href='/vendor/basscss.min.css' />
-                <link rel='stylesheet' type='text/css' href='/vendor/font-awesome-4.7.0/css/font-awesome.min.css' />
+                <link rel='stylesheet' type='text/css' href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css' />
                 <link rel='stylesheet' type='text/css' href='/vendor/video-js.min.css' />
                 <div style="display:none;">Server in ${region}</div>
               </body>